+++ deside/workflow/workflow.py
@@ -3,7 +3,7 @@ import numpy as np
 import pandas as pd
 import warnings
 from ..decon_cf import DeSide
-from ..utility import check_dir, print_msg, set_fig_style
+from ..utility import check_dir, print_msg
 from ..utility.read_file import ReadH5AD
 from ..utility.compare import mean_exp_of_marker_gene, read_and_merge_result, cal_gene_signature_score
 from ..plot import (compare_exp_and_cell_fraction, plot_predicted_result,
@@ -15,7 +15,7 @@ warnings.simplefilter(action='ignore', category=UserWarning)
 def tcga_evaluation(marker_gene_file_path, total_result_dir, pred_cell_frac_tcga_dir,
                     cell_types, tcga_data_dir, outlier_file_path=None, pre_trained_model_dir=None,
                     model_name: str = None, signature_score_method: str = 'mean_exp', cancer_types: list = None,
-                    update_figures=False):
+                    update_figures=False, pathway_mask=None):
     """
 
     :param marker_gene_file_path:
@@ -31,6 +31,7 @@ def tcga_evaluation(marker_gene_file_path, total_result_dir, pred_cell_frac_tcga
         gene_signature_score (Combes et al., 2022, Cell 185, 184-203)
     :param cancer_types: a list of cancer types
     :param update_figures: update figures or not
+    :param pathway_mask: genes by pathways, 1 for genes in the pathway, 0 for genes not in the pathway
     :return:
     """
     # marker_gene_file_path = marker_gene_file_path
@@ -49,7 +50,6 @@ def tcga_evaluation(marker_gene_file_path, total_result_dir, pred_cell_frac_tcga
     signature_score_result_dir = os.path.join(pred_cell_frac_dir_current_model, 'signature_score')
     check_dir(signature_score_result_dir)
     all_signature_score_file_path = os.path.join(signature_score_result_dir, 'all_cancer_type_signature_score.csv')
-    # cell_types = cell_types
     if 'DeSide' in model_name:
         if os.path.exists(os.path.join(pre_trained_model_dir, model_name, model_name, 'genes.txt')):
             gene_list_in_model_fp = os.path.join(pre_trained_model_dir, model_name, model_name, 'genes.txt')
@@ -95,10 +95,6 @@ def tcga_evaluation(marker_gene_file_path, total_result_dir, pred_cell_frac_tcga
     # combine all predicted cell fraction for each cancer type together
     if not os.path.exists(all_pred_cell_frac_file_path):
         print(f'   Merge all predicted result by {model_name}...')
-        # if 'DeSide' in model:
-        #     _cell_types = cell_types + ['1-others']
-        # else:
-        #     _cell_types = cell_types
         _cell_type_name_mapping = dict(zip(cell_types, cell_types))
         read_and_merge_result(raw_result_dir=pred_cell_frac_dir_current_model,
                               cell_type_name_mapping=_cell_type_name_mapping,
@@ -120,12 +116,13 @@ def tcga_evaluation(marker_gene_file_path, total_result_dir, pred_cell_frac_tcga
     # outlier_file_path = 'outlier_samples.txt'
     outlier_file_path = outlier_file_path
 
-    cell_types_clustering = [i for i in cell_types if i != 'Cancer Cells']
-    compare_exp_and_cell_fraction(merged_file_path=merged_signature_score_and_cell_frac_file_path,
-                                  clustering_ct=cell_types_clustering, font_scale=1.5,
-                                  cell_types=cell_types, outlier_file_path=outlier_file_path,
-                                  result_dir=result_dir_new, update_figures=update_figures,
-                                  signature_score_method=signature_score_method)
+    if pathway_mask is None:
+        cell_types_clustering = [i for i in cell_types if i != 'Cancer Cells']
+        compare_exp_and_cell_fraction(merged_file_path=merged_signature_score_and_cell_frac_file_path,
+                                      clustering_ct=cell_types_clustering, font_scale=1.5,
+                                      cell_types=cell_types, outlier_file_path=outlier_file_path,
+                                      result_dir=result_dir_new, update_figures=update_figures,
+                                      signature_score_method=signature_score_method)
 
     print('Plot predicted cell proportion across all cancer types...')
     cell_types2max = {'B Cells': 0.1, 'CD4 T': 0.1, 'DC': 0.1, 'CD8 T': 0.1}
@@ -140,7 +137,20 @@ def tcga_evaluation(marker_gene_file_path, total_result_dir, pred_cell_frac_tcga
 
 
 def run_step3(evaluation_dataset2path, log_file_path, result_dir, model_dir,
-              all_cell_types, one_minus_alpha=False):
+              all_cell_types, one_minus_alpha=False, pathway_mask=None,
+              method_adding_pathway='add_to_end', hyper_params: dict = None):
+    """
+    Step3: Predicting cell fractions of test set and evaluation
+    :param evaluation_dataset2path: dict, key: dataset name, value: file path
+    :param log_file_path: str, log file path
+    :param result_dir: str, result directory
+    :param model_dir: str, model directory
+    :param all_cell_types: list, all cell types
+    :param one_minus_alpha: bool, whether to use 1-alpha as the predicted cell fraction for all cell types
+    :param pathway_mask: dataframe, pathway mask, genes by pathways
+    :param method_adding_pathway: str, method for adding pathway, 'add_to_end' or 'convert'
+    :param hyper_params: dict, hyper parameters for DNN model
+    """
     # Step3, evaluation on test set
     print_msg('Step3: Predicting cell fractions of test set and evaluation...',
               log_file_path=log_file_path)
@@ -164,7 +174,9 @@ def run_step3(evaluation_dataset2path, log_file_path, result_dir, model_dir,
             deside_model.predict(input_file=generated_bulk_gep_fp,
                                  output_file_path=predicted_cell_frac_file_path,
                                  exp_type='log_space', scaling_by_sample=False,
-                                 scaling_by_constant=True, one_minus_alpha=one_minus_alpha)
+                                 scaling_by_constant=True, one_minus_alpha=one_minus_alpha,
+                                 pathway_mask=pathway_mask, method_adding_pathway=method_adding_pathway,
+                                 hyper_params=hyper_params)
         print('   > Comparing cell frac between y_true and y_pred...')
         for cell_type in generated_cell_frac.columns.to_list():
             s_plot = ScatterPlot(x=predicted_cell_frac_file_path,
@@ -184,10 +196,32 @@ def run_step3(evaluation_dataset2path, log_file_path, result_dir, model_dir,
                               y_label=f'y_pred')
 
 
-def run_step4(tcga_data_dir, cancer_types, log_file_path, model_dir, marker_gene_file_path,
-              result_dir, pred_cell_frac_tcga_dir, cancer_purity_file_path, all_cell_types,
-              model_names, signature_score_method, one_minus_alpha=False,
-              update_figures=False, outlier_file_path=None):
+def run_step4(tcga_data_dir: str, cancer_types: list, log_file_path: str, model_dir: str,
+              marker_gene_file_path: str, result_dir: str, pred_cell_frac_tcga_dir: str,
+              cancer_purity_file_path: str, all_cell_types: list, model_names: list,
+              signature_score_method: str, one_minus_alpha: bool = False,
+              update_figures: bool = False, outlier_file_path: str = None, pathway_mask: pd.DataFrame = None,
+              method_adding_pathway: str = 'add_to_end', hyper_params: dict = None):
+    """
+    Step4: Predicting cell fractions of TCGA
+    :param tcga_data_dir: str, TCGA data directory
+    :param cancer_types: list, cancer types
+    :param log_file_path: str, log file path
+    :param model_dir: str, model directory
+    :param marker_gene_file_path: str, marker gene file path
+    :param result_dir: str, result directory
+    :param pred_cell_frac_tcga_dir: str, predicted cell fraction of TCGA directory
+    :param cancer_purity_file_path: str, cancer purity file path
+    :param all_cell_types: list, all cell types
+    :param model_names: list, model names
+    :param signature_score_method: str, signature score method
+    :param one_minus_alpha: bool, whether to use 1-alpha as the predicted cell fraction for all cell types
+    :param update_figures: bool, whether to update figures
+    :param outlier_file_path: str, outlier file path
+    :param pathway_mask: dataframe, pathway mask, genes by pathways
+    :param method_adding_pathway: str, method for adding pathway, 'add_to_end' or 'convert'
+    :param hyper_params: dict, hyper parameters for DNN model
+    """
     # TCGA
     print_msg("Step 4: Predict cell fraction of TCGA...", log_file_path=log_file_path)
     # model_name = 'DeSide'
@@ -206,7 +240,9 @@ def run_step4(tcga_data_dir, cancer_types, log_file_path, model_dir, marker_gene
                 deside_model = DeSide(model_dir=model_dir)
                 deside_model.predict(input_file=current_bulk_tpm, output_file_path=y_pred_file_path,
                                      exp_type='TPM', scaling_by_constant=True,
-                                     scaling_by_sample=False, one_minus_alpha=one_minus_alpha)
+                                     scaling_by_sample=False, one_minus_alpha=one_minus_alpha,
+                                     pathway_mask=pathway_mask, method_adding_pathway=method_adding_pathway,
+                                     hyper_params=hyper_params)
             else:
                 print(f'   Previous result existed: {y_pred_file_path}')
             print(f'   Plot and compare predicted result...')
@@ -220,7 +256,8 @@ def run_step4(tcga_data_dir, cancer_types, log_file_path, model_dir, marker_gene
                         cell_types=all_cell_types, tcga_data_dir=tcga_data_dir,
                         pre_trained_model_dir=model_dir, model_name=model_name,
                         signature_score_method=signature_score_method, cancer_types=cancer_types,
-                        update_figures=update_figures, outlier_file_path=outlier_file_path)
+                        update_figures=update_figures, outlier_file_path=outlier_file_path,
+                        pathway_mask=pathway_mask)
 
         # calculate the distribution of predicted cell proportions in TCGA
         # model_name = 'DeSide'